Simple table from chapter 8 (Creating a Simple Table Based App)
