+++ SimpleTable/SimpleTable/Assets.xcassets/AccentColor.colorset/Contents.json
