+++ SimpleTable/SimpleTable/Assets.xcassets/AppIcon.appiconset/Contents.json
