+++ SimpleTable/SimpleTable/Assets.xcassets/Contents.json
