+++ SimpleTable/SimpleTable/Info.plist
