+++ SimpleTable/SimpleTable/ViewController.swift
@@ -1,0 +1,31 @@
+//
+//  ViewController.swift
+//  SimpleTable
+//
+//  Created by Viviana Mesaros on 07.04.2022.
+//
+
+import UIKit
+
+class ViewController: UIViewController,UITableViewDataSource, UITableViewDelegate {
+
+    override var prefersStatusBarHidden: Bool {
+        return true
+    }
+
+    var restaurantNames = ["Cafe Deadend", "Homei", "Teakha", "Cafe Loisl", "Petite Oyster", "For Kee Restaurant", "Po's Atelier", "Bourke Street Bakery", "Haigh's Chocolate", "Palomino Espresso", "Upstate", "Traif", "Graham Avenue Meats And Deli", "Waffle & Wolf", "Five Leaves", "Cafe Lore", "Confessional", "Barrafina", "Donostia", "Royal Oak", "CASK Pub and Kitchen"]
+    
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return restaurantNames.count
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cellIdentifier = "datacell"
+        let cell = tableView.dequeueReusableCell(withIdentifier: cellIdentifier, for: indexPath)
+        
+        cell.textLabel?.text = restaurantNames[indexPath.row]
+        cell.imageView?.image = UIImage(named: "restaurant")
+        
+        return cell
+    }
+}
