--- chapter6/exercise2/exercise2/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  exercise2
-//
-//  Created by Viviana Mesaros on 06.04.2022.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
