--- chapter6/exercises3/exercises3/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  exercises3
-//
-//  Created by Viviana Mesaros on 07.04.2022.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
